Add user_id to items table
In the case of an airbnb type place an item (listing) will only ever
belong to one user.  This relationship allows us to scope listings based
on user (for example when a host wants to modify their listings, they
shouldn't be permitted to edit any listings other than their own.