Add helper to return collection of dates from a range
We are creating an entry in the availabilities table for each date in
the range specified by a user.  This method takes two args, start and
end date, and returns an array of hashes, each hash meant to be passed
to create whenever a host creates a new item or a user adds an item to
an order (i.e. whenever we need to store a range of available dates).
Data structure returned looks like this:
  [
    {date: <DateObject>},
    {date: <DateObject>}
  ]

With the assumption that this return val will be passed directly into
something like item.availabilities.create