Add datepicker to items#new (fixes #79)
A few issues here: how to get capybara to select item from the
datepicker.  That also ties into how we want to validate for dates on
an item (are there any scenarios where we would have an item without a
set of availabilities?).

As it is now an exception is raised when you try to create/update an
item with no start/end date provided.