Always verify both user_id and remember_token when checking current user
If both of these properties are relevant to us being (securely) sure that the current user has not been spoofed, then we should always be verifying for both of these. Either being incorrect would indicate a tampered session token. We should not be conditionally returning 404 in this case, we should be erroring out as soon as humanly possible.  is the logical place for this, though an argument could be made for setting current_user to nil if the remember_token doesn't match, in the case that they signed in on another machine. This should be investigated later.