Fix tests
I got rid of a bunch of obsolete tests, and did minor patchwork where I
could.  The few that are left pending are either related to things we
may need in the future (addresses, cc info, etc), or ones that were
tricky enough to make me want to leave them for next week.